Fix controllers

+++ green-barrel/src/controllers/auto_slug.rs
@@ -23,7 +23,7 @@ pub struct AutoSlug {
     pub hint: String,        // Additional explanation for the user.
     pub warning: String,     // The value is determined automatically.
     pub error: String,       // The value is determined automatically.
-    pub group: u64, // To optimize field traversal in the `paladins/check()` method. Hint: It is recommended not to change.
+    pub group: u32, // To optimize field traversal in the `paladins/check()` method. Hint: It is recommended not to change.
 }
 
 impl Default for AutoSlug {
@@ -47,7 +47,7 @@ impl Default for AutoSlug {
             hint: String::new(),
             warning: String::new(),
             error: String::new(),
-            group: 2,
+            group: 2_u32,
         }
     }
 }
