+++ green-barrel/src/controllers/number_i64.rs
@@ -26,7 +26,7 @@ pub struct NumberI64 {
     pub hint: String,        // Additional explanation for the user.
     pub warning: String,     // The value is determined automatically.
     pub error: String,       // The value is determined automatically.
-    pub group: u64, // To optimize field traversal in the `paladins/check()` method. Hint: It is recommended not to change.
+    pub group: u32, // To optimize field traversal in the `paladins/check()` method. Hint: It is recommended not to change.
 }
 
 impl Default for NumberI64 {
@@ -53,7 +53,7 @@ impl Default for NumberI64 {
             hint: String::new(),
             warning: String::new(),
             error: String::new(),
-            group: 11,
+            group: 11_u32,
         }
     }
 }
