+++ green-barrel/src/controllers/number_u32.rs
@@ -10,23 +10,23 @@ pub struct NumberU32 {
     pub field_type: String, // Field type.
     pub input_type: String, // The value is determined automatically.
     pub name: String, // The value is determined automatically.
-    pub value: Option<i64>, // Sets the value of an element.
-    pub default: Option<i64>, // Value by default.
+    pub value: Option<u32>, // Sets the value of an element.
+    pub default: Option<u32>, // Value by default.
     pub placeholder: String, // Displays prompt text.
     pub required: bool, // Mandatory field.
     pub unique: bool, // The unique value of a field in a collection.
     pub disabled: bool, // Blocks access and modification of the element.
     pub readonly: bool, // Specifies that the field cannot be modified by the user.
-    pub step: i64,  // Increment step for numeric fields.
-    pub min: Option<i64>, // The lower value for entering a number or date.
-    pub max: Option<i64>, // The top value for entering a number or date.
+    pub step: u32,  // Increment step for numeric fields.
+    pub min: Option<u32>, // The lower value for entering a number or date.
+    pub max: Option<u32>, // The top value for entering a number or date.
     pub is_hide: bool, // Hide field from user.
     pub other_attrs: String, // Example: r# "autofocus tabindex="some number" size="some numberString::new()#.
     pub css_classes: String, // Example: "class-name-1 class-name-2".
     pub hint: String,        // Additional explanation for the user.
     pub warning: String,     // The value is determined automatically.
     pub error: String,       // The value is determined automatically.
-    pub group: u64, // To optimize field traversal in the `paladins/check()` method. Hint: It is recommended not to change.
+    pub group: u32, // To optimize field traversal in the `paladins/check()` method. Hint: It is recommended not to change.
 }
 
 impl Default for NumberU32 {
@@ -44,7 +44,7 @@ impl Default for NumberU32 {
             unique: false,
             disabled: false,
             readonly: false,
-            step: 1,
+            step: 1_u32,
             min: None,
             max: None,
             is_hide: false,
@@ -53,13 +53,13 @@ impl Default for NumberU32 {
             hint: String::new(),
             warning: String::new(),
             error: String::new(),
-            group: 11,
+            group: 11_u32,
         }
     }
 }
 
 impl NumberU32 {
     pub fn set(&mut self, value: u32) {
-        self.value = Some(i64::from(value));
+        self.value = Some(value);
     }
 }
