+++ green-barrel/src/controllers/radio_u32.rs
@@ -10,20 +10,20 @@ pub struct RadioU32 {
     pub field_type: String, // Field type.
     pub input_type: String, // The value is determined automatically.
     pub name: String, // The value is determined automatically.
-    pub value: Option<i64>, // Sets the value of an element.
-    pub default: Option<i64>, // Value by default.
+    pub value: Option<u32>, // Sets the value of an element.
+    pub default: Option<u32>, // Value by default.
     pub placeholder: String, // Displays prompt text.
     pub required: bool, // Mandatory field.
     pub disabled: bool, // Blocks access and modification of the element.
     pub readonly: bool, // Specifies that the field cannot be modified by the user.
-    pub options: Vec<(i64, String)>, // Html tag: <option value="5">Title</option> ; Example: vec![(5, "Title"), (25, "Title 2")].
+    pub options: Vec<(u32, String)>, // Html tag: <option value="5">Title</option> ; Example: vec![(5, "Title"), (25, "Title 2")].
     pub is_hide: bool,               // Hide field from user.
     pub other_attrs: String, // Example: r# "autofocus tabindex="some number" size="some numberString::new()#.
     pub css_classes: String, // Example: "class-name-1 class-name-2".
     pub hint: String,        // Additional explanation for the user.
     pub warning: String,     // The value is determined automatically.
     pub error: String,       // The value is determined automatically.
-    pub group: u64, // To optimize field traversal in the `paladins/check()` method. Hint: It is recommended not to change.
+    pub group: u32, // To optimize field traversal in the `paladins/check()` method. Hint: It is recommended not to change.
 }
 
 impl Default for RadioU32 {
@@ -47,13 +47,13 @@ impl Default for RadioU32 {
             hint: String::new(),
             warning: String::new(),
             error: String::new(),
-            group: 11,
+            group: 11_u32,
         }
     }
 }
 
 impl RadioU32 {
     pub fn set(&mut self, value: u32) {
-        self.value = Some(i64::from(value));
+        self.value = Some(value);
     }
 }
