+++ green-barrel/src/controllers/select_i32_mult.rs
@@ -25,7 +25,7 @@ pub struct SelectI32Mult {
     pub hint: String,        // Additional explanation for the user.
     pub warning: String,     // The value is determined automatically.
     pub error: String,       // The value is determined automatically.
-    pub group: u64, // To optimize field traversal in the `paladins/check()` method. Hint: It is recommended not to change.
+    pub group: u32, // To optimize field traversal in the `paladins/check()` method. Hint: It is recommended not to change.
 }
 
 impl Default for SelectI32Mult {
@@ -51,7 +51,7 @@ impl Default for SelectI32Mult {
             hint: String::new(),
             warning: String::new(),
             error: String::new(),
-            group: 6,
+            group: 6_u32,
         }
     }
 }
