+++ green-barrel/src/controllers/select_i32_mult_dyn.rs
@@ -24,7 +24,7 @@ pub struct SelectI32MultDyn {
     pub hint: String,        // Additional explanation for the user.
     pub warning: String,     // The value is determined automatically.
     pub error: String,       // The value is determined automatically.
-    pub group: u64, // To optimize field traversal in the `paladins/check()` method. Hint: It is recommended not to change.
+    pub group: u32, // To optimize field traversal in the `paladins/check()` method. Hint: It is recommended not to change.
 }
 
 impl Default for SelectI32MultDyn {
@@ -49,7 +49,7 @@ impl Default for SelectI32MultDyn {
             hint: String::new(),
             warning: String::new(),
             error: String::new(),
-            group: 7,
+            group: 7_u32,
         }
     }
 }
