+++ green-barrel/src/controllers/select_text.rs
@@ -26,7 +26,7 @@ pub struct SelectText {
     pub hint: String,        // Additional explanation for the user.
     pub warning: String,     // The value is determined automatically.
     pub error: String,       // The value is determined automatically.
-    pub group: u64, // To optimize field traversal in the `paladins/check()` method. Hint: It is recommended not to change.
+    pub group: u32, // To optimize field traversal in the `paladins/check()` method. Hint: It is recommended not to change.
 }
 
 impl Default for SelectText {
@@ -40,8 +40,8 @@ impl Default for SelectText {
             value: None,
             default: None,
             placeholder: String::new(),
-            minlength: 0,
-            maxlength: 256,
+            minlength: 0_usize,
+            maxlength: 256_usize,
             required: false,
             unique: false,
             disabled: false,
@@ -53,7 +53,7 @@ impl Default for SelectText {
             hint: String::new(),
             warning: String::new(),
             error: String::new(),
-            group: 4,
+            group: 4_u32,
         }
     }
 }
