+++ green-barrel/src/controllers/select_text_dyn.rs
@@ -25,7 +25,7 @@ pub struct SelectTextDyn {
     pub hint: String,        // Additional explanation for the user.
     pub warning: String,     // The value is determined automatically.
     pub error: String,       // The value is determined automatically.
-    pub group: u64, // To optimize field traversal in the `paladins/check()` method. Hint: It is recommended not to change.
+    pub group: u32, // To optimize field traversal in the `paladins/check()` method. Hint: It is recommended not to change.
 }
 
 impl Default for SelectTextDyn {
@@ -38,8 +38,8 @@ impl Default for SelectTextDyn {
             name: String::new(),
             value: None,
             placeholder: String::new(),
-            minlength: 0,
-            maxlength: 256,
+            minlength: 0_usize,
+            maxlength: 256_usize,
             required: false,
             unique: false,
             disabled: false,
@@ -51,7 +51,7 @@ impl Default for SelectTextDyn {
             hint: String::new(),
             warning: String::new(),
             error: String::new(),
-            group: 5,
+            group: 5_u32,
         }
     }
 }
