+++ green-barrel/src/controllers/select_text_mult.rs
@@ -26,7 +26,7 @@ pub struct SelectTextMult {
     pub hint: String,        // Additional explanation for the user.
     pub warning: String,     // The value is determined automatically.
     pub error: String,       // The value is determined automatically.
-    pub group: u64, // To optimize field traversal in the `paladins/check()` method. Hint: It is recommended not to change.
+    pub group: u32, // To optimize field traversal in the `paladins/check()` method. Hint: It is recommended not to change.
 }
 
 impl Default for SelectTextMult {
@@ -40,8 +40,8 @@ impl Default for SelectTextMult {
             value: None,
             default: None,
             placeholder: String::new(),
-            minlength: 0,
-            maxlength: 256,
+            minlength: 0_usize,
+            maxlength: 256_usize,
             required: false,
             unique: false,
             disabled: false,
@@ -53,7 +53,7 @@ impl Default for SelectTextMult {
             hint: String::new(),
             warning: String::new(),
             error: String::new(),
-            group: 6,
+            group: 6_u32,
         }
     }
 }
