+++ green-barrel/src/controllers/select_text_mult_dyn.rs
@@ -25,7 +25,7 @@ pub struct SelectTextMultDyn {
     pub hint: String,        // Additional explanation for the user.
     pub warning: String,     // The value is determined automatically.
     pub error: String,       // The value is determined automatically.
-    pub group: u64, // To optimize field traversal in the `paladins/check()` method. Hint: It is recommended not to change.
+    pub group: u32, // To optimize field traversal in the `paladins/check()` method. Hint: It is recommended not to change.
 }
 
 impl Default for SelectTextMultDyn {
@@ -38,8 +38,8 @@ impl Default for SelectTextMultDyn {
             name: String::new(),
             value: None,
             placeholder: String::new(),
-            minlength: 0,
-            maxlength: 256,
+            minlength: 0_usize,
+            maxlength: 256_usize,
             required: false,
             unique: false,
             disabled: false,
@@ -51,7 +51,7 @@ impl Default for SelectTextMultDyn {
             hint: String::new(),
             warning: String::new(),
             error: String::new(),
-            group: 7,
+            group: 7_u32,
         }
     }
 }
